Process tag files using multiprocessing pools
With huge tag files, such as the ones you get from the Linux kernel,
parsing all of the tags and filtering with fnmatch can be a very
processer extensive task and take a while. Splitting this into multiple
processes (since threads are useless for this in python) makes this
exponentially faster.

The unfortunate downside is some accuracy is lost on the progress
indicator, but this isn't super important.