generate.py: Implement actual status indicators for workers
The current solution is slow since in order to continually update the
status indicator on the terminal, we have to keep the chunksize down to
5000 which slows things down on systems with a lot of cores due to the
constant overhead of sending/receiving objects.

In order to work around the lack of built-in status indicator in pools,
we create a simple array in shared memory (with locking disabled) that's
the length of the number of workers we have. All workers are initialized
with a shared Value, and each worker locks said Value, stores the
integer in it, and increments it, resulting in each worker having it's
own designated spot in the array. From then on, each worker stores the
number of items it's processed in the array, and the main thread
periodically loops through the array and adds the values all up.

This isn't prone to race conditions since even though the shared array
has no lock, there's only ever one process writing to it at a time. As
well, this allows us to get a much more accurate indication of how much
work has been done. Plus, we can make the chunksize as large as we want
now.