Introduce abstract GeneratorBase class
This is a class that will now act as the base class for any syntax
generator. As well, the class is used during the tag parsing process so
that language specific syntax generators can hook additional code into
the filtering phase of tag parsing, allowing us to filter out as many
tags as possible that we're not going to use in the highlight object
population phase. As well, this makes the API that syntax generators
must implement a lot more clear.