syntax.py: Add generate_init_script(), fix relative paths
For some reason, @% isn't actually always guaranteed to be a relative
path. This can break TagBlock()'s scope comparison script and as a
result, cause all local tags in a file not to be highlighted as a result
of @% having a different value then the expected filepath from the tag
file.

To fix this, we add a generate_init_script() function for syntax.py so
that we can setup non-filetype specific global state at the start of the
script. Using this, we add the variable s:rel_file into our generated
vimscript, which corrects absolute @% paths referring to files in the
project directory to relative paths.